LPD-75170 refactor: rename ai-hub-api to ai-hub-nexus-api
