--- modules/dxp/apps/ai-hub-nexus/ai-hub-api/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay AI Hub API
-Bundle-SymbolicName: com.liferay.ai.hub.api
-Bundle-Version: 2.0.0
-Export-Package:\
-	com.liferay.ai.hub.agent,\
-	com.liferay.ai.hub.util
