+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-api/src/main/java/com/liferay/ai/hub/nexus/agent/AgentContext.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.hub.agent;
+package com.liferay.ai.hub.nexus.agent;
 
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.vulcan.dto.converter.DTOConverterContext;
