+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-api/src/main/resources/com/liferay/ai/hub/nexus/agent/packageinfo
