+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/java/com/liferay/ai/hub/nexus/web/internal/display/context/EditAgentDefinitionDisplayContext.java
@@ -7,7 +7,7 @@ package com.liferay.ai.hub.nexus.web.internal.display.context;
 
 import com.liferay.account.model.AccountEntry;
 import com.liferay.account.service.AccountEntryLocalService;
-import com.liferay.ai.hub.util.AccountEntryUtil;
+import com.liferay.ai.hub.nexus.util.AccountEntryUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Company;
