+++ modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/MessageResourceImpl.java
@@ -5,8 +5,8 @@
 
 package com.liferay.ai.hub.rest.internal.resource.v1_0;
 
-import com.liferay.ai.hub.agent.AgentContext;
-import com.liferay.ai.hub.agent.SupervisorAgent;
+import com.liferay.ai.hub.nexus.agent.AgentContext;
+import com.liferay.ai.hub.nexus.agent.SupervisorAgent;
 import com.liferay.ai.hub.rest.dto.v1_0.Message;
 import com.liferay.ai.hub.rest.internal.resource.v1_0.util.GroupUtil;
 import com.liferay.ai.hub.rest.resource.v1_0.MessageResource;
